move sim_fixed
