+++ sim/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(easy)
 add_subdirectory(exp_window)
-# add_subdirectory(fixed)
+add_subdirectory(fixed)
 # add_subdirectory(sliding_window)
